chore: fixed daysOfSchool routes
