+++ src/models/attendances.ts
@@ -13,6 +13,16 @@ const attendances = new Schema({
     minlength: 12,
     maxlength: 12,
   },
+  academicYear: {
+    from: String,
+    to: String,
+    required: true,
+  },
+  gradeLevel: {
+    type: String,
+    enum: ['1', '2', '3', '4', '5', '6', '7', '8', '9', '10'],
+    required: true,
+  },
   createdAt: {
     type: Date,
     default: Date.now,
