+++ src/routes/index.ts
@@ -1,7 +1,7 @@
 import { Application } from 'express'
 import { API_ROOT } from '../utils/constants'
-import SchoolDaysRoutes from './schoolDays'
+import DaysOfSchoolRoutes from './daysOfSchool'
 
 export default function (app: Application) {
-  app.use(`${API_ROOT}/schooldays`, SchoolDaysRoutes)
+  app.use(`${API_ROOT}/days-of-school`, DaysOfSchoolRoutes)
 }
